requirements.txt
pyyaml
slackclient
tqdm
emoji_data_python
pillow
